Inject ConsumedAndProducedLocations
