+++ subprojects/core/src/main/java/org/gradle/execution/plan/LocalTaskNode.java
@@ -16,8 +16,6 @@
 
 package org.gradle.execution.plan;
 
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.ImmutableSortedMap;
 import org.gradle.api.Action;
 import org.gradle.api.Task;
 import org.gradle.api.internal.TaskInternal;
@@ -30,7 +28,6 @@ import org.gradle.api.internal.tasks.properties.PropertyWalker;
 import org.gradle.api.internal.tasks.properties.TaskProperties;
 import org.gradle.api.tasks.TaskExecutionException;
 import org.gradle.internal.ImmutableActionSet;
-import org.gradle.internal.fingerprint.CurrentFileCollectionFingerprint;
 import org.gradle.internal.resources.ResourceDeadlockException;
 import org.gradle.internal.resources.ResourceLock;
 import org.gradle.internal.service.ServiceRegistry;
@@ -38,7 +35,6 @@ import org.gradle.internal.service.ServiceRegistry;
 import javax.annotation.Nullable;
 import java.io.File;
 import java.util.List;
-import java.util.Optional;
 import java.util.Set;
 
 /**
@@ -51,8 +47,6 @@ public class LocalTaskNode extends TaskNode {
     private boolean finalized;
     private List<? extends ResourceLock> resourceLocks;
     private TaskProperties taskProperties;
-    private ImmutableSet<String> inputFileLocations;
-    private Runnable validationAction;
 
     public LocalTaskNode(TaskInternal task) {
         this.task = task;
@@ -276,25 +270,5 @@ public class LocalTaskNode extends TaskNode {
         } catch (ResourceDeadlockException e) {
             throw new IllegalStateException(String.format("A deadlock was detected while resolving the %s for task '%s'. This can be caused, for instance, by %s property causing dependency resolution.", description, task, singular), e);
         }
-    }
-
-    public void setInputs(ImmutableSortedMap<String, CurrentFileCollectionFingerprint> inputFileFingerprints) {
-        ImmutableSet.Builder<String> builder = ImmutableSet.builder();
-        for (CurrentFileCollectionFingerprint fingerprint : inputFileFingerprints.values()) {
-            builder.addAll(fingerprint.getRootPaths());
-        }
-        inputFileLocations = builder.build();
-    }
-
-    public Optional<ImmutableSet<String>> getInputFileLocations() {
-        return Optional.ofNullable(inputFileLocations);
-    }
-
-    public Runnable getValidationAction() {
-        return validationAction;
-    }
-
-    public void setValidationAction(Runnable validationAction) {
-        this.validationAction = validationAction;
     }
 }
